Replace 'char' type by 'uint8_t' type in saved structures. Needed by the Lua-based save/load system. Adapt code accordingly.
git-svn-id: https://freedroid.svn.sourceforge.net/svnroot/freedroid@4727 524c2bfb-fa3f-0410-b5a2-92f4e1dc5b61

+++ src/lua.c
@@ -818,7 +818,7 @@ static int lua_set_bot_destination(lua_State * L)
 
 static int lua_set_rush_tux(lua_State * L)
 {
-	const int cmd = luaL_checkinteger(L, 1);
+	const uint8_t cmd = (uint8_t)luaL_checkinteger(L, 1);
 	enemy *en = get_enemy_arg(L, 2);
 	en->will_rush_tux = cmd;
 	return 0;
